New Cool Stuff!!!
- New command ".among" has finally been added! Staff members can now give the Among Us role to members without asking to another staff member.

- The bot is user-end is now 100% in Portuguese!

- New enums to make the code cleaner.

- Uptime command is now working as the expected.

- Default messages: Some messages are not store in `Mesages.class` enum, just to standardize some messages.

- The word "idade" in the messages sent in the console "Menor de Idade" and "Maior de Idade" should start with lowercase (fixed now).

TODO: Finish `/disconnect` slash command.

+++ src/main/java/bot/commands/ColorRoleSchedule.java
@@ -75,6 +75,8 @@ public class ColorRoleSchedule {
 
         if (logChannel != null) logChannel.sendMessage(logMessage).queue();
         else System.out.println("Não foi possível encontrar o canal de registros para cargo de cor, ignorando...");
+
+
     }
 
     private static Member fetchTarget(String arg, Guild guild) {
@@ -89,23 +91,23 @@ public class ColorRoleSchedule {
 
         switch (arg) {
             case "fire", "fireelement" -> {
-                return guild.getRoleById(ColorRoles.FIRE_ELEMENT);
+                return guild.getRoleById(ColorRoles.FIRE_ELEMENT.toLong());
             }
 
             case "earth", "earthelement" -> {
-                return guild.getRoleById(ColorRoles.EARTH_ELEMENT);
+                return guild.getRoleById(ColorRoles.EARTH_ELEMENT.toLong());
             }
 
             case "water", "waterelement" -> {
-                return guild.getRoleById(ColorRoles.WATER_ELEMENT);
+                return guild.getRoleById(ColorRoles.WATER_ELEMENT.toLong());
             }
 
             case "light", "lightelement" -> {
-                return guild.getRoleById(ColorRoles.LIGHT_ELEMENT);
+                return guild.getRoleById(ColorRoles.LIGHT_ELEMENT.toLong());
             }
 
             case "air", "airelement" -> {
-                return guild.getRoleById(ColorRoles.AIR_ELEMENT);
+                return guild.getRoleById(ColorRoles.AIR_ELEMENT.toLong());
             }
 
             default -> {
